feat(api): Add /issues/<group_id>/events/<event_id> endpoint
