+++ tests/sentry/api/endpoints/test_group_event_details.py
@@ -4,7 +4,7 @@ from sentry.testutils.silo import region_silo_test
 
 
 @region_silo_test(stable=True)
-class GroupEventsOldestEndpointTest(APITestCase, SnubaTestCase):
+class GroupEventDetailsEndpointTest(APITestCase, SnubaTestCase):
     def setUp(self):
         super().setUp()
 
@@ -39,7 +39,16 @@ class GroupEventsOldestEndpointTest(APITestCase, SnubaTestCase):
             project_id=project.id,
         )
 
-    def test_get_simple(self):
+    def test_get_simple_latest(self):
+        url = f"/api/0/issues/{self.event_a.group.id}/events/latest/"
+        response = self.client.get(url, format="json")
+
+        assert response.status_code == 200, response.content
+        assert response.data["id"] == str(self.event_c.event_id)
+        assert response.data["previousEventID"] == str(self.event_b.event_id)
+        assert response.data["nextEventID"] is None
+
+    def test_get_simple_oldest(self):
         url = f"/api/0/issues/{self.event_a.group.id}/events/oldest/"
         response = self.client.get(url, format="json")
 
@@ -48,7 +57,25 @@ class GroupEventsOldestEndpointTest(APITestCase, SnubaTestCase):
         assert response.data["previousEventID"] is None
         assert response.data["nextEventID"] == str(self.event_b.event_id)
 
-    def test_get_with_environment(self):
+    def test_get_simple_event_id(self):
+        url = f"/api/0/issues/{self.event_b.group.id}/events/{self.event_b.event_id}/"
+        response = self.client.get(url, format="json")
+
+        assert response.status_code == 200, response.content
+        assert response.data["id"] == str(self.event_b.event_id)
+        assert response.data["previousEventID"] == str(self.event_a.event_id)
+        assert response.data["nextEventID"] == str(self.event_c.event_id)
+
+    def test_get_with_environment_latest(self):
+        url = f"/api/0/issues/{self.event_a.group.id}/events/latest/"
+        response = self.client.get(url, format="json", data={"environment": ["production"]})
+
+        assert response.status_code == 200, response.content
+        assert response.data["id"] == str(self.event_b.event_id)
+        assert response.data["previousEventID"] is None
+        assert response.data["nextEventID"] is None
+
+    def test_get_with_environment_oldest(self):
         url = f"/api/0/issues/{self.event_a.group.id}/events/oldest/"
         response = self.client.get(url, format="json", data={"environment": ["production"]})
 
@@ -56,3 +83,12 @@ class GroupEventsOldestEndpointTest(APITestCase, SnubaTestCase):
         assert response.data["id"] == str(self.event_b.event_id)
         assert response.data["previousEventID"] is None
         assert response.data["nextEventID"] is None
+
+    def test_collapse_stacktrace_only(self):
+        url = f"/api/0/issues/{self.event_a.group.id}/events/latest/"
+        response = self.client.get(url, format="json", data={"collapse": ["stacktraceOnly"]})
+
+        assert response.status_code == 200, response.content
+        assert response.data["id"] == str(self.event_c.event_id)
+        assert "previousEventID" not in response.data
+        assert "nextEventID" not in response.data
